+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## TBD
+## v48.1.0 (2023-03-27)
 
 ### Added